+++ mobile/src/login/LoginForm.tests.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import Providers from 'TestProviders';
+
+import LoginForm from './';
+
+describe('The LoginForm component', () => {
+  it('should render', () => {
+    const tree = renderer.create(
+      <Providers>
+        <LoginForm />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
+});