+++ mobile/src/tabs/connect/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { ScrollView } from 'react-native';
 import { createStackNavigator } from 'react-navigation';
+import LoginForm from 'login';
 import BackgroundView from 'ui/BackgroundView';
 import TableView, { Cell, CellIcon, CellText, Divider } from 'ui/TableView';
 import tabBarIcon from '../tabBarIcon';
@@ -38,6 +39,7 @@ export class ConnectScreen extends React.Component {
               <CellIcon name="building" />
               <CellText>Find Service Times & Locations</CellText>
             </Cell>
+            <LoginForm />
           </TableView>
         </ScrollView>
       </BackgroundView>